fix: 修复审查发现的 6 项 bug 并补齐回归测试
覆盖代码审查定位的全部可复现问题：

- adapter-trimmer 参数校验：拒绝 adapter-min-overlap=0 与
  max-mismatches >= min-overlap（后者使重叠比对退化为必然命中，
  每个 read 的 3' 端被静默截断）
- stdin gzip 嗅探：循环凑满 2 字节魔数并重试 EINTR，
  慢管道单字节短读不再误报 FORMAT 错误
- stat 输出目标：output/json/signature 任一组合中多个 '-' 同时
  写 stdout 会被拒绝（JSON 后被拼接 TSV 导致不可解析）
- writer 原子发布：rename 前 fsync 数据页并 fsync 父目录，
  掉电场景下不再可能发布空/截断文件（best-effort）
- signature-kmer-size=0 钳制为 1，消除报告空 key 行
- 质量字符统一按 int8_t 解释，x86/ARM 对 >=128 非法字节行为一致，
  且与 AVX2 有符号比较语义对齐

TDD：全部修复先写失败测试（单测/e2e）再实现；-funsigned-char
对照验证跨平台一致性。

+++ src/processing/predicates/min_quality_predicate.cpp
@@ -1,5 +1,7 @@
 #include "fqtools/processing/predicates/min_quality_predicate.h"
 
+#include <cstdint>
+
 #include <fmt/format.h>
 
 namespace fq::processing {
@@ -33,7 +35,8 @@ auto MinQualityPredicate::calculateAverageQuality(std::string_view qualityString
     // 导致坏数据（ASCII < encoding）反而通过质控
     int64_t sumQual = 0;
     for (char q : qualityString) {
-        const int qVal = static_cast<int>(q) - qualityEncoding_;
+        // 显式按 int8_t 解释：非法字节（>=128）在 x86/ARM 上一致视为负质量
+        const int qVal = static_cast<int>(static_cast<std::int8_t>(q)) - qualityEncoding_;
         sumQual += (qVal > 0) ? qVal : 0;
     }
     return static_cast<double>(sumQual) / static_cast<double>(qualityString.size());
